разделение редьюсеров и экшенов по файлам
