+++ src/components/IngredientDetails/IngredientDetails.jsx
@@ -5,7 +5,7 @@ import styles from './ingredientDetails.module.css'
 import {useSelector} from "react-redux";
 
 function IngredientDetails () {
-    const item = useSelector(state => state.itemReducer.item);
+    const item = useSelector(state => state.modal.item);
 return(
     <div >
        <h1 className="text text_type_main-large mt-10 ml-10">
